+++ packages/aws-appsync-nodejs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.9.19 (2026-07-03)
+
+**Note:** Version bump only for package @ttoss/aws-appsync-nodejs
+
 ## 1.9.18 (2026-06-10)
 
 **Note:** Version bump only for package @ttoss/aws-appsync-nodejs